+++ lib/src/interfaces/interfaces.dart
@@ -1,17 +1,17 @@
 library interfaces;
 
-import 'dart:convert';
 import 'dart:typed_data';
 
 import 'package:passkeys/types.dart';
+import 'package:web3_signers/src/vendor/vendor.dart' show U8aExtension;
 import 'package:web3dart/crypto.dart';
 import 'package:web3dart/web3dart.dart';
 
-import '../utils/utils.dart' show Uint256, hexStripPrefix;
+import '../utils/utils.dart' show Uint256, hexlify;
 import '../web3_signers_base.dart' show PassKeyPair, PassKeySignature;
 
 part 'eoa_wallet_interface.dart';
 part 'multi_signer_interface.dart';
 part 'passkey_signer_interface.dart';
-part 'uint256_interface.dart';
 part 'signature_options.dart';
+part 'uint256_interface.dart';