+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.3
+
+- seperate passkey build signature bytes
+- use abi.encode instead instead of manual encoding
+- Use Uint8list instead of strings
+
 ## 0.1.2
 
 - optimize signature options